feat/ovos_audio_compat_+_autoconfigure
